split the draw plugin logic in two different files
